requirements.txt
asgiref==3.9.1
Django==5.2.6
djangorestframework==3.16.1
dotenv==0.9.9
et_xmlfile==2.0.0
openpyxl==3.1.5
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.3
typing_extensions==4.15.0
